remove entities and definition tab components
