--- packages/hash/frontend/src/pages/[account-slug]/types/entity-type/entity-type-entities-tab.tsx
@@ -1,20 +1,0 @@
-import { TabProps } from "@mui/material";
-import { FunctionComponent } from "react";
-import { useEntityTypeEntities } from "./use-entity-type-entities";
-import { TabLink } from "./tab-link";
-
-export const EntityTypeEntitiesTab: FunctionComponent<TabProps> = ({
-  value,
-  ...props
-}) => {
-  const { entities } = useEntityTypeEntities() ?? {};
-
-  return (
-    <TabLink
-      {...props}
-      value={value}
-      label="Entities"
-      count={entities?.length}
-    />
-  );
-};
